pyproject.toml
[project]
name = "trello-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "fastmcp==2.13.1",
    "pydantic-settings>=2.12.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.12.0",
]
